+++ test/unit/index.test.js
@@ -96,5 +96,33 @@ describe('manager', () => {
             expect(env.test.paths).to.equal('development');
         });
     });
+
+    context('when passed cli arguments', () => {
+        it('should override config values', () => {
+            const env = manager({
+                dir: path.resolve(__dirname, '../fixtures/file-based'),
+                argv: ['--env', 'dev', '--paths.output', 'lib']
+            });
+
+            expect(env).to.exist;
+            expect(env.paths.output).to.equal('lib');
+        });
+
+        it('should override config values including base ones', () => {
+            const env = manager({
+                pattern: '{platform}/{env}.json',
+                dir: path.resolve(__dirname, '../fixtures/directory-based'),
+                argv: ['--env', 'dev', '--platform', 'linux', '--test.server.port', 3030],
+                base: 'environment.json'
+            });
+
+            expect(env).to.exist;
+            expect(env.build).to.exist;
+            expect(env.build.watch).to.equal(true);
+            expect(env.test.server.port).to.equal(3030);
+            expect(env.test.server.singleRun).to.false;
+            expect(env.test.paths).to.equal('dev');
+        });
+    });
 });
 /* eslint-enable no-unused-expressions */